+++ src/lib/api.js
@@ -198,6 +198,32 @@ const validateStartAndEndDatetimes = function (startDateTime, endDateTime) {
   }
 }
 
+/**
+ * ensure that fields necessary for creating links i.e. 'collection' and 'id'
+ * are not excluded from query.  These fields will be removed later to ensure
+ * results match user expectations
+ * @param {Object} fields
+ * @returns {Object}
+ */
+export const createQueryFields = function (fields) {
+  const { exclude } = fields
+  if (exclude) {
+    const filteredExclude = exclude.filter(
+      (field) => field !== 'id' && field !== 'collection'
+    )
+    if (filteredExclude.length === 0) {
+      const { exclude: _removed, ...rest } = fields
+      return rest
+    }
+
+    return {
+      ...fields,
+      exclude: filteredExclude
+    }
+  }
+  return fields
+}
+
 export const extractDatetime = function (params) {
   const { datetime } = params
 
@@ -540,7 +566,6 @@ export const addItemLinks = function (results, endpoint) {
   results.forEach((result) => {
     let { links } = result
     const { id, collection } = result
-
     links = (links === undefined) ? [] : links
     // self link
     links.splice(0, 0, {
@@ -573,6 +598,32 @@ export const addItemLinks = function (results, endpoint) {
     }
     result.type = 'Feature'
     return result
+  })
+  return results
+}
+
+/**
+ * If 'id' or 'collection' were in the 'excluded' fields, they must
+ * be removed.  They were necessary for STAC Item link generation and
+ * can now be removed after link generation if a user wanted to exclude them
+ * Impure, we are potentially mutating 'results'
+ * @param {Object} results
+ * @param {Object} fields - {'exclude': [string], 'include': [string]}
+ * @returns {Object}
+ */
+export const removeSpecialExcludeFields = function (results, fields) {
+  const { exclude } = fields
+  if (!exclude) return results
+
+  const removeId = exclude.includes('id')
+  const removeCollection = exclude.includes('collection')
+
+  // exit early and avoid forEach loop if possible
+  if (!removeId && !removeCollection) return results
+
+  results.forEach((item) => {
+    if (removeId) delete item.id
+    if (removeCollection) delete item.collection
   })
   return results
 }
@@ -688,6 +739,7 @@ const searchItems = async function (
     extractRestrictionCql2Filter(parameters, headers)
   )
   const fields = extractFields(parameters)
+  const queryFields = createQueryFields(fields)
   const ids = extractIds(parameters)
   const allowedCollectionIds = extractAllowedCollectionIds(
     parameters,
@@ -704,7 +756,7 @@ const searchItems = async function (
     query,
     filter: combinedFilter,
     sortby,
-    fields,
+    fields: queryFields,
     ids,
     collections,
     next
@@ -798,6 +850,7 @@ const searchItems = async function (
   }
 
   addItemLinks(responseItems, endpoint)
+  removeSpecialExcludeFields(responseItems, fields)
 
   return wrapResponseInFeatureCollection(responseItems, links, numberMatched, numberReturned, limit)
 }